Get predictions working with concode and non-trained model
The loop at run.py:378 seems to be the "core loop" for prediction. One iteration generates one
character of the output. The whole string of [input] + [generated output]
must never get longer than 1024, otherwise the model chokes because it's hardcoded
(in its config.json on huggingfaces) to accept only input strings of length
1024.

This was causing problems because there are NL prompts in the test set that
are over 900 tokens long and the code in run.py generates up to 100 tokens of output.
The way the original code dealt with this was to set block_size to 512 when loading the dataset.
This means, as far as I can tell, that all NL prompts are truncated to at most 512 tokens.
I had set block_size to 1024 because it's the default in the original argparser code.